Always display inferior process buffer after evaluation
Previously, if the inferior process buffer is not being shown in any
window, one would have to manually bring it up after evaluating an
expression in order to see the results. With this change, the inferior
process buffer will always be displayed after evaluating an expression
so that one can alway see the results without having to bring up the
inferior process buffer.